+++ lib/cli/src/commands/compile.rs
@@ -87,7 +87,7 @@ impl Compile {
                 if ext != recommended_extension {
                     warning!("the output file has a wrong extension. We recommend using `{}.{}` for the chosen target", &output_filename, &recommended_extension)
                 }
-            },
+            }
             None => {
                 warning!("the output file has no extension. We recommend using `{}.{}` for the chosen target", &output_filename, &recommended_extension)
             }