+++ lib/wasi/src/lib.rs
@@ -30,6 +30,8 @@ pub use crate::utils::{get_wasi_version, is_wasi_module, WasiVersion};
 
 use thiserror::Error;
 use wasmer::{imports, Function, ImportObject, Memory, Module, Store};
+#[cfg(all(target_os = "macos", target_arch = "aarch64",))]
+use wasmer::{FunctionType, ValType};
 
 use std::cell::UnsafeCell;
 use std::fmt;
@@ -50,7 +52,9 @@ pub enum WasiError {
 /// The environment provided to the WASI imports.
 #[derive(Debug, Clone)]
 pub struct WasiEnv {
-    state: Arc<Mutex<WasiState>>,
+    /// Shared state of the WASI system. Manages all the data that the
+    /// executing WASI program can see.
+    pub state: Arc<Mutex<WasiState>>,
     memory: Arc<WasiMemory>,
 }
 
@@ -174,7 +178,7 @@ impl WasiEnv {
     }
 
     pub(crate) fn get_memory_and_wasi_state(
-        &mut self,
+        &self,
         _mem_index: u32,
     ) -> (&Memory, MutexGuard<WasiState>) {
         let memory = self.memory();
@@ -197,6 +201,34 @@ pub fn generate_import_object_from_env(
             generate_import_object_snapshot1(store, wasi_env)
         }
     }
+}
+
+// Note: we use this wrapper because native functions with more than 9 params
+// fail on Apple Silicon (with Cranelift).
+fn get_path_open_for_store(store: &Store, env: WasiEnv) -> Function {
+    #[cfg(not(all(target_os = "macos", target_arch = "aarch64",)))]
+    let path_open = Function::new_native_with_env(store, env.clone(), path_open);
+    #[cfg(all(target_os = "macos", target_arch = "aarch64",))]
+    let path_open = Function::new_with_env(
+        store,
+        &FunctionType::new(
+            vec![
+                ValType::I32,
+                ValType::I32,
+                ValType::I32,
+                ValType::I32,
+                ValType::I32,
+                ValType::I64,
+                ValType::I64,
+                ValType::I32,
+                ValType::I32,
+            ],
+            vec![ValType::I32],
+        ),
+        env.clone(),
+        path_open_dynamic,
+    );
+    path_open
 }
 
 /// Combines a state generating function with the import list for legacy WASI
@@ -234,7 +266,7 @@ fn generate_import_object_snapshot0(store: &Store, env: WasiEnv) -> ImportObject
             "path_filestat_get" => Function::new_native_with_env(store, env.clone(), legacy::snapshot0::path_filestat_get),
             "path_filestat_set_times" => Function::new_native_with_env(store, env.clone(), path_filestat_set_times),
             "path_link" => Function::new_native_with_env(store, env.clone(), path_link),
-            "path_open" => Function::new_native_with_env(store, env.clone(), path_open),
+            "path_open" => get_path_open_for_store(store, env.clone()),
             "path_readlink" => Function::new_native_with_env(store, env.clone(), path_readlink),
             "path_remove_directory" => Function::new_native_with_env(store, env.clone(), path_remove_directory),
             "path_rename" => Function::new_native_with_env(store, env.clone(), path_rename),
@@ -287,7 +319,7 @@ fn generate_import_object_snapshot1(store: &Store, env: WasiEnv) -> ImportObject
             "path_filestat_get" => Function::new_native_with_env(store, env.clone(), path_filestat_get),
             "path_filestat_set_times" => Function::new_native_with_env(store, env.clone(), path_filestat_set_times),
             "path_link" => Function::new_native_with_env(store, env.clone(), path_link),
-            "path_open" => Function::new_native_with_env(store, env.clone(), path_open),
+            "path_open" => get_path_open_for_store(store, env.clone()),
             "path_readlink" => Function::new_native_with_env(store, env.clone(), path_readlink),
             "path_remove_directory" => Function::new_native_with_env(store, env.clone(), path_remove_directory),
             "path_rename" => Function::new_native_with_env(store, env.clone(), path_rename),