+++ lib/wasi/src/ptr.rs
@@ -24,6 +24,12 @@ impl<T: Copy, Ty> fmt::Debug for WasmPtr<T, Ty> {
     }
 }
 
+impl<T: Copy, Ty> From<i32> for WasmPtr<T, Ty> {
+    fn from(offset: i32) -> Self {
+        Self::new(offset as _)
+    }
+}
+
 unsafe impl<T: Copy, Ty> FromToNativeWasmType for WasmPtr<T, Ty> {
     type Native = <BaseWasmPtr<T, Ty> as FromToNativeWasmType>::Native;